add loading data

+++ service/src/main/java/com/softserveinc/trainee/dao/Impl/EntityDaoImpl.java
@@ -2,15 +2,19 @@ package com.softserveinc.trainee.dao.Impl;
 
 import com.softserveinc.trainee.dao.EntityDao;
 import com.softserveinc.trainee.entity.metadata.Entity;
+import com.softserveinc.trainee.entity.metadata.Field;
 import org.springframework.stereotype.Repository;
 
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.persistence.Query;
+import javax.ws.rs.ClientErrorException;
+import javax.ws.rs.core.Response;
+import java.util.Date;
 import java.util.List;
 
 @Repository("entityDao")
-public class EntityDaoImpl implements EntityDao{
+public class EntityDaoImpl implements EntityDao {
 
     @PersistenceContext(unitName = "application")
     private EntityManager entityManager;
@@ -27,18 +31,30 @@ public class EntityDaoImpl implements EntityDao{
 
     @Override
     public Entity updateEntity(Entity entity) {
-        /*Entity persistEntity = entityManager.find(Entity.class, entity.getId());
+        Entity persistEntity = entityManager.find(Entity.class, entity.getId());
         if(persistEntity != null){
-            entityManager.remove(persistEntity);
+            entity.setCreatedDate(persistEntity.getCreatedDate());
+            for(Field fieldPersisted: persistEntity.getFieldList()){
+                for(Field field: entity.getFieldList()){
+                    if(fieldPersisted.getId().equals(field.getId())){
+                        field.setCreatedDate(fieldPersisted.getCreatedDate());
+                    }
+                }
+            }
+            entityManager.merge(entity);
+            return entity;
         }
-        entityManager.persist(entity);*/
-        entityManager.merge(entity);
-        return entity;
+        throw new ClientErrorException(Response.Status.NOT_FOUND);
     }
 
     @Override
-    public void deleteEntity(Entity entity) {
+    public void deleteEntity(String id) {
+        Entity entity = entityManager.find(Entity.class, id);
+        if(entity != null){
             entityManager.remove(entity);
+        }else{
+            throw new ClientErrorException(Response.Status.NOT_FOUND);
+        }
     }
 
     @Override
